Move 404 test page to applications (and tidy up)
